Cast millis into f32 when converting dt into milliseconds.

+++ src/game.rs
@@ -108,13 +108,13 @@ impl Game {
             return;
         }
 
-        let millis = dt.as_millis();
+        let millis = dt.as_millis() as f32;
 
-        self.right_paddle.pos.Y += self.right_player.movement * PADDLE_VELOCITY * millis as f32;
-        self.left_paddle.pos.Y += self.left_player.movement * PADDLE_VELOCITY * millis as f32;
+        self.right_paddle.pos.Y += self.right_player.movement * PADDLE_VELOCITY * millis;
+        self.left_paddle.pos.Y += self.left_player.movement * PADDLE_VELOCITY * millis;
 
-        self.ball.pos.Y += self.ball_y_movement * BALL_VELOCITY * millis as f32;
-        self.ball.pos.X += self.ball_x_movement * BALL_VELOCITY * millis as f32;
+        self.ball.pos.Y += self.ball_y_movement * BALL_VELOCITY * millis;
+        self.ball.pos.X += self.ball_x_movement * BALL_VELOCITY * millis;
 
         // reflect ball Y-movement if it hits the bottom wall.
         if self.bottom_wall.pos.Y <= (self.ball.pos.Y + 0.0325) {
